backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
google-cloud-firestore>=2.18.0
google-cloud-storage>=2.18.0
google-genai>=1.35.0
